Extract some tests about datagram reading
